Реализован переход от EnsureCreated к MigrateAsync
