+++ Models/DailyReadingData.cs
@@ -1,0 +1,12 @@
+namespace Library.Models
+{
+    /// <summary>
+    /// Данные о чтении за один день
+    /// </summary>
+    public class DailyReadingData
+    {
+        public DateTime Date { get; set; }
+        public int PagesRead { get; set; }
+    }
+}
+
